requirements.txt
appnope @ file:///Users/builder/ci_310/appnope_1642500616005/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi==2022.5.18.1
charset-normalizer==2.0.12
cycler==0.11.0
debugpy @ file:///Users/builder/ci_310/debugpy_1642501698574/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
entrypoints @ file:///opt/concourse/worker/volumes/live/5eb4850e-dcbc-41ad-5f22-922bac778f70/volume/entrypoints_1649926457041/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fonttools==4.33.3
idna==3.3
ipykernel @ file:///opt/concourse/worker/volumes/live/b899b150-2bf0-46bf-5730-9e8a3e9ee440/volume/ipykernel_1647000787125/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_2ff86d21-7133-4fee-b3b2-6349ba52809eua87k1km/croots/recipe/ipython_1651600160735/work
jedi @ file:///opt/concourse/worker/volumes/live/18b71546-5bde-4add-72d1-7d16b76f0f7a/volume/jedi_1644315243726/work
joblib==1.1.0
jupyter-client @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_4d0a367d-74c5-447f-831e-3e1fd3984797bfrg56pj/croots/recipe/jupyter_client_1650622206454/work
jupyter-core @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_8ba802e0-15f6-4a45-a9ee-144034526783uwxtvvew/croots/recipe/jupyter_core_1651671235924/work
kiwisolver==1.4.2
matplotlib==3.5.2
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
nest-asyncio @ file:///opt/concourse/worker/volumes/live/4c28d7d6-d544-4f26-7811-eac0cb2a490e/volume/nest-asyncio_1649847935212/work
numpy==1.22.4
packaging==21.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing==3.0.9
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pyzmq @ file:///Users/builder/ci_310/pyzmq_1642536888228/work
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
threadpoolctl==3.1.0
torch==1.11.0
torchvision==0.12.0
tornado @ file:///Users/builder/ci_310/tornado_1642538510757/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
